Check whether snap is installed at all, and catalog presence
Throw a `SnapError` if `snapd` isn't installed, to provide more
meaningful feedback to users.

`snapd` does not always populate the catalog in
`/var/cache/snapd/names` at install time, especially if it's
installed/started as part of `cloud-init`, in a container, or
somewhere else where networking may not be available at first
run.

The snapd API doesn't provide a convenient way to refresh this,
and restarting the service is undesirable. Instead, if a requested
package isn't in the cache, blindly try the `snap` API to see if
we can fetch info for the requested package and return a
`SnapNotFoundError` if we get a `SnapAPIError`