Added systemd.py
This is mainly a port from charmhelpers, focused on routines that
allow charm authors to stop, start, and otherwise control systemd
managed services.

This code is still a draft. It needs a proper testing strategy, and
also possibly some simplifications (this library has to do less than
the original charmhelpers host.py did).