Fix failing test.
One of our unit tests was reaching out to the system to do a fairly
harmless query about a snap name.

On systems without snapd installed (e.g., a mac, or a github
container), this would fail. So we mock out the call to check_output,
and all is good!