[Draft] Added snap cohorts
Addresses gh#9

This was meant to be a quick "dive in a fix it."

We haven't actually implemented refresh, however, which means that
there is more work to do.

Next steps are to add enough code to fix the failing tests, while giving
some thought to how refresh should flow through the library in general.