Systemd library now raises Exceptions in place of returning bools
Per the request in GH#33, we refactored the systemd lib to raise
Exceptions when it fails to do something, rather than returning false.

A path in the code keeps the behavior of is_running consistent. On a
return code of 3, it will return False for "is not running", rather
than raising an Exception.