Systemd library now raises Exceptions in place of returning bools (#36)
Per the request in GH#33, we refactored the systemd lib to raise
Exceptions when it fails to do something, rather than returning false.

A branch in the logic keeps the behavior of is_running consistent. On a
return code of 3, it will return False for "is not running", rather
than raising an Exception.

This PR also simplifies the behavior of service_pause and service_resume. Since systemd behaves idempotently, and doesn't complain if a user repeatedly masks or unmasks a service, we don't need to routinely squelch errors when doing so.