passwd: create home dir for system users
The "--system" argument needs also "--create-home" argument to be
present for the creation of home directory for system users.

This change adds a new parameter to the add_user function signature,
called `create_home` which defaults to True to provide backwards
compatibility with the current api. After this change, home directory
will be always created regardless of regular users or system users
unless False is passed for `create_home`.

Fixes: #52
Signed-off-by: Mert Kırpıcı <mert.kirpici@canonical.com>