fix!: add support for observing snap services
Adds the `watch.yaml` file for configuring how the notices
daemon observes services. The daemon will get the service
alias from this file and use the alias to determine the
proper hook to fire when the observed service is started or stopped.

BREAKING CHANGES: __init__ for SystemdNotices now takes the Service
data class instead of a generic list. The Service dataclass was added
so that it would be easier to specify an alias for an observed service.

Signed-off-by: Jason C. Nucciarone <nuccitheboss@ubuntu.com>