ci: update linting (following ops) (#143)
This PR updates `tox.ini` to use `ruff --preview` for all linting and formatting (dropping `black`), and have `codespell` run correctly, updating `pyproject.toml` accordingly.

The core of the PR is the changes to `tox.ini` and `pyproject.toml`. The linting rules for `ruff` have been updated to include all the rules that `ops` uses. To allow linting to continue to pass, any rules that are currently broken in a single file are ignored for that file, while rules that are broken in multiple files are ignored globally. These ignores are marked with a `FIXME` to distinguish them from intentional ignores that are copied from `ops`.

Changes to other files are either from `ruff format` or from `codespell`. There are typo corrections and a couple of small formatting changes to libraries, which are not significant enough to warrant bumping `LIB_PATCH` and triggering a release. There are also some bigger formatting changes for large dictionaries in tests.

Subsequent PRs will fix the ignored linting issues in individual libraries.