style: use type annotations from the future (for apt.py) (#149)
The PR updates the `apt` library to use futurised type annotations. This will make it easier to read the diff for the next PR, which will update type annotations for this library to pass pyright's type checking. `LIBPATCH` will be incremented once this batch of PRs for the `apt` library are completed.