feat: improve logging on snap error (#162)
When a `SnapError` is raised after a `snap` command errors (raising a `CalledProcessError`), the error does not currently contain the process's `stderr`. Additionally, the failure information may not be present in `stdout` or `stderr`, but may instead have been logged to the journal. This PR updates the 6 call sites where a `CalledProcessError` is converted to a `SnapError` as follows:
- The `subprocess` call is updated to always capture `stdout` and `stderr` so that these are available in the `CalledProcessError`
- The specific error message and the `CalledProcessError` are passed to a new constructor for `SnapError`
- `SnapError._from_called_process_error` extends the provided message with `stdout` and `stderr` from the `CalledProcessError`, and attempts to extend it further with the last 20 entries from the journal from the `snapd` unit -- if this fails, the failure is appended to the message instead

The unit tests have been updated due to the changes to the `subprocess` calls. The functional change is exercised in an integration test, which has been updated to be more strict (previously it could pass even if the expected error wasn't raised), and to check that the logs have been retrieved successfully.